wobbly: implement translation command
This is necessary when for ex. a view is moved between two outputs with
Expo active. The view geometry then changes across multiple workspaces,
and we cannot just move the grab because the view drags behind.

Instead, we translate the whole view when such a thing happens, ensuring
smooth dragging to matter where the view is and which plugins are
active.